Fixes banner to from date:time save action
When magento is configured for a non-usa locale, strtotime will always return false if the slash character is used to separate d/m/y. User complaints regarding dates not saving in admin will all have non-usa locales in common.
This fix replaces the slash with a period if locale is non-usa.